pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "recipeparser"
version = "4.0.0"
description = "Extract recipes from EPUB cookbooks and export to Paprika 3"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "Ian Ballard"}
]
dependencies = [
    "EbookLib>=0.18",
    "beautifulsoup4>=4.12.2",
    "lxml>=5.3.1",
    "pydantic>=2.11.9",
    "google-genai>=1.38.0",
    "python-dotenv>=1.1.1",
    "pyyaml>=6.0.1",
    "customtkinter>=5.2.2",
    "pymupdf>=1.24.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.30.0",
    "httpx>=0.27.0",
    "PyJWT>=2.8.0",
]

[project.scripts]
recipeparser = "recipeparser.__main__:main"
recipeparser-gui = "recipeparser.gui:run_gui"

[tool.setuptools.packages.find]
where = ["."]
include = ["recipeparser*"]

[tool.setuptools.package-data]
recipeparser = ["*.yaml"]

[tool.pytest.ini_options]
# live_*_test.py files are standalone scripts that require a running server.
# They are NOT pytest suites — exclude them from the standard test run.
# collect_ignore is handled in conftest.py (the correct location for this setting).
testpaths = ["tests"]
python_files = ["test_*.py"]
